test: split test_tools_execution_input_gates.py into 6 per-concern files
Split the 810-line test_tools_execution_input_gates.py (12 test classes plus
9 module-level helpers) into per-concern files plus a shared helper module:

- _input_contract_test_helpers.py (new, ~50 LoC): centralized helpers consumed
  by 3 of the 6 split files:
    _FixedUUID, _DETERMINISTIC_HEX, _DETERMINISTIC_MARKER, _patch_uuid4,
    _make_input_contract_resolver, _collect_all_path_input_specs,
    _collect_file_path_list_specs, _ALL_PATH_INPUT_SPECS, _FILE_PATH_LIST_SPECS
- test_tools_execution_input_gates_dry_walkthrough.py (~190 LoC):
  TestCheckDryWalkthrough + TestDryWalkthroughGateWithPrefix
- test_tools_execution_input_gates_command.py (~135 LoC):
  TestRunSkillPrefix + TestValidateSkillCommand
- test_tools_execution_input_gates_contracts.py (~140 LoC):
  TestInputContractValidation + TestInputContractResolver +
  TestInputContractIntegration (uses shared helpers)
- test_tools_execution_input_gates_real_contracts.py (~145 LoC):
  TestInputContractRealContracts + TestFilePathListRealResolver (uses shared
  _ALL_PATH_INPUT_SPECS / _FILE_PATH_LIST_SPECS)
- test_tools_execution_input_gates_cross_binding.py (~95 LoC):
  TestPreparePrCrossBinding + TestOpenIntegrationPrExactShape
- test_tools_execution_input_gates_cwd.py (~50 LoC):
  TestRunSkillCwdValidation (uses shared _patch_uuid4 / _DETERMINISTIC_MARKER)

Mechanical lift: no production code changes, no behavior changes, no test
assertion changes. Each file carries pytestmark = [layer("server"), small].
The shared helper module centralizes the YAML-loaded contract-spec tables
(loaded once at import time rather than per test class) and the deterministic
UUID fixtures used by the CWD and integration tests.

Refs: #4610
