--- tests/server/test_tools_execution_input_gates.py
@@ -1,810 +1,0 @@
-"""Tests for run_skill input validation gates and CWD checking."""
-
-from __future__ import annotations
-
-import json
-from pathlib import Path
-from typing import TYPE_CHECKING, cast
-
-if TYPE_CHECKING:
-    from typing import Literal
-
-import pytest
-
-from autoskillit.server._guards import _check_dry_walkthrough
-from autoskillit.server._state import _get_config
-from autoskillit.server.tools.tools_execution import run_skill
-
-pytestmark = [pytest.mark.layer("server"), pytest.mark.small]
-
-# Deterministic UUID for tests that need to predict the per-invocation marker.
-_DETERMINISTIC_HEX = "a1b2c3d4e5f6a7b890123456"
-_DETERMINISTIC_MARKER = f"%%ORDER_UP::{_DETERMINISTIC_HEX[:8]}%%"
-
-
-class _FixedUUID:
-    hex = _DETERMINISTIC_HEX
-
-
-def _patch_uuid4(monkeypatch):
-    """Monkeypatch uuid4 to return a deterministic value for marker prediction."""
-    monkeypatch.setattr("uuid.uuid4", lambda: _FixedUUID())
-
-
-class TestCheckDryWalkthrough:
-    """Dry-walkthrough gate blocks both /implement-worktree variants."""
-
-    def test_dry_walkthrough_gate_blocks_implement_no_merge(self, tool_ctx, tmp_path):
-        """Gate blocks /implement-worktree-no-merge when plan lacks marker."""
-        plan = tmp_path / "plan.md"
-        plan.write_text("# My Plan\n\nSome content")
-        result = _check_dry_walkthrough(f"/implement-worktree-no-merge {plan}", str(tmp_path))
-        assert result is not None
-        parsed = json.loads(result)
-        assert parsed["success"] is False
-        assert parsed["is_error"] is True
-        assert "dry-walked" in parsed["result"].lower()
-
-    def test_dry_walkthrough_gate_passes_implement_no_merge(self, tool_ctx, tmp_path):
-        """Gate allows /implement-worktree-no-merge when plan has marker."""
-        plan_dir = tmp_path / "make-plan"
-        plan_dir.mkdir()
-        plan = plan_dir / "plan.md"
-        plan.write_text("Dry-walkthrough verified = TRUE\n# My Plan")
-        result = _check_dry_walkthrough(f"/implement-worktree-no-merge {plan}", str(tmp_path))
-        assert result is None
-
-    def test_dry_walkthrough_gate_still_works_for_implement_worktree(self, tool_ctx, tmp_path):
-        """Original /implement-worktree gating is not broken."""
-        plan = tmp_path / "plan.md"
-        plan.write_text("# No marker plan")
-        result = _check_dry_walkthrough(f"/implement-worktree {plan}", str(tmp_path))
-        assert result is not None
-        parsed = json.loads(result)
-        assert parsed["success"] is False
-        assert parsed["is_error"] is True
-
-    def test_dry_walkthrough_gate_ignores_unrelated_skills(self, tool_ctx):
-        """Gate ignores skills that are not implement-worktree variants."""
-        result = _check_dry_walkthrough("/autoskillit:investigate some-error", "/tmp")
-        assert result is None
-
-    def test_dry_walkthrough_gate_with_part_a_named_file_marked(self, tmp_path, tool_ctx):
-        """Gate accepts _part_a.md file when marker is present."""
-        plan_dir = tmp_path / "make-plan"
-        plan_dir.mkdir()
-        plan = plan_dir / "task_plan_2026-01-01_part_a.md"
-        plan.write_text("Dry-walkthrough verified = TRUE\n\nContent here")
-        result = _check_dry_walkthrough(f"/implement-worktree-no-merge {plan}", str(tmp_path))
-        assert result is None
-
-    def test_dry_walkthrough_gate_with_part_b_named_file_unmarked(self, tmp_path, tool_ctx):
-        """Gate blocks _part_b.md file when marker is absent."""
-        plan = tmp_path / "task_plan_2026-01-01_part_b.md"
-        plan.write_text("> **PART B ONLY.**\n\nNo walkthrough marker here")
-        result = _check_dry_walkthrough(f"/implement-worktree-no-merge {plan}", str(tmp_path))
-        assert result is not None
-        parsed = json.loads(result)
-        assert parsed["subtype"] == "gate_error"
-
-    def test_dry_walkthrough_gate_distinguishes_parts_independently(self, tmp_path, tool_ctx):
-        """Gate correctly distinguishes marked part_a from unmarked part_b."""
-        plan_dir = tmp_path / "make-plan"
-        plan_dir.mkdir()
-        part_a = plan_dir / "task_plan_part_a.md"
-        part_b = tmp_path / "task_plan_part_b.md"
-        part_a.write_text("Dry-walkthrough verified = TRUE\n\nPart A content")
-        part_b.write_text("> **PART B ONLY.**\n\nPart B content — no marker")
-
-        result_a = _check_dry_walkthrough(f"/implement-worktree-no-merge {part_a}", str(tmp_path))
-        result_b = _check_dry_walkthrough(f"/implement-worktree-no-merge {part_b}", str(tmp_path))
-        assert result_a is None
-        assert result_b is not None
-        assert json.loads(result_b)["subtype"] == "gate_error"
-
-    def test_gate_with_trailing_markdown_header_finds_plan(self, tmp_path, tool_ctx):
-        """Trailing markdown headers must not corrupt the plan path."""
-        plan_dir = tmp_path / "make-plan"
-        plan_dir.mkdir()
-        plan = plan_dir / "plan.md"
-        plan.write_text(_get_config().implement_gate.marker + "\n\nrest")
-        cmd = f"/implement-worktree-no-merge {plan}\n\n## Base Branch\nimpl-926"
-        assert _check_dry_walkthrough(cmd, str(tmp_path)) is None
-
-    def test_gate_with_extra_token_after_path(self, tmp_path, tool_ctx):
-        """Space-separated token after path must not corrupt the plan path."""
-        plan_dir = tmp_path / "make-plan"
-        plan_dir.mkdir()
-        plan = plan_dir / "plan.md"
-        plan.write_text(_get_config().implement_gate.marker + "\n\nrest")
-        cmd = f"/implement-worktree-no-merge {plan} impl-926"
-        assert _check_dry_walkthrough(cmd, str(tmp_path)) is None
-
-    def test_gate_multiline_no_marker_reports_dry_walk_error(self, tmp_path, tool_ctx):
-        """With trailing headers and plan missing marker: dry-walk error, not file-not-found."""
-        plan = tmp_path / "plan.md"
-        plan.write_text("# Plan title\n\nNo marker here")
-        cmd = f"/implement-worktree-no-merge {plan}\n\n## Base Branch\nimpl-926"
-        result = _check_dry_walkthrough(cmd, str(tmp_path))
-        assert result is not None
-        data = json.loads(result)
-        message = data.get("result", "").lower()
-        assert "not found" not in message, "Should fail on marker absence, not path lookup"
-        assert "dry-walk" in message or "dry-walked" in message
-
-    # --- Plan-origin (allowed_plan_dirs) validation tests ---
-
-    def test_dry_walkthrough_gate_blocks_plan_at_wrong_directory(self, tool_ctx, tmp_path):
-        """Gate blocks plan file not in an allowed origin directory."""
-        wrong_dir = tmp_path / "dry-walkthrough"
-        wrong_dir.mkdir()
-        plan = wrong_dir / "plan.md"
-        plan.write_text("Dry-walkthrough verified = TRUE\n# Plan")
-        result = _check_dry_walkthrough(f"/implement-worktree-no-merge {plan}", str(tmp_path))
-        assert result is not None
-        parsed = json.loads(result)
-        assert parsed["success"] is False
-        assert (
-            "original location" in parsed["result"].lower()
-            or "allowed" in parsed["result"].lower()
-        )
-
-    def test_dry_walkthrough_gate_allows_plan_at_make_plan_dir(self, tool_ctx, tmp_path):
-        """Gate allows plan file in the make-plan directory."""
-        plan_dir = tmp_path / "make-plan"
-        plan_dir.mkdir()
-        plan = plan_dir / "plan.md"
-        plan.write_text("Dry-walkthrough verified = TRUE\n# Plan")
-        result = _check_dry_walkthrough(f"/implement-worktree-no-merge {plan}", str(tmp_path))
-        assert result is None
-
-    def test_dry_walkthrough_gate_allows_plan_at_rectify_dir(self, tool_ctx, tmp_path):
-        """Gate allows plan file in the rectify directory."""
-        plan_dir = tmp_path / "rectify"
-        plan_dir.mkdir()
-        plan = plan_dir / "plan.md"
-        plan.write_text("Dry-walkthrough verified = TRUE\n# Plan")
-        result = _check_dry_walkthrough(f"/implement-worktree-no-merge {plan}", str(tmp_path))
-        assert result is None
-
-    def test_dry_walkthrough_gate_blocks_plan_at_arbitrary_dir(self, tool_ctx, tmp_path):
-        """Gate blocks plan at arbitrary non-origin directory."""
-        arb_dir = tmp_path / "audit-impl"
-        arb_dir.mkdir()
-        plan = arb_dir / "plan.md"
-        plan.write_text("Dry-walkthrough verified = TRUE\n# Plan")
-        result = _check_dry_walkthrough(f"/implement-worktree-no-merge {plan}", str(tmp_path))
-        assert result is not None
-        parsed = json.loads(result)
-        assert parsed["subtype"] == "gate_error"
-
-    def test_implement_gate_config_default_allowed_dirs(self, tool_ctx):
-        """Default allowed_plan_dirs includes make-plan and rectify."""
-        from autoskillit.server._state import _get_config
-
-        config = _get_config()
-        assert "make-plan" in config.implement_gate.allowed_plan_dirs
-        assert "rectify" in config.implement_gate.allowed_plan_dirs
-
-
-class TestRunSkillPrefix:
-    """run_skill passes prefixed command to subprocess."""
-
-    @pytest.mark.anyio
-    async def test_run_skill_prefixes_skill_command(self, tool_ctx_kitchen_open):
-        from tests.conftest import _make_result
-
-        tool_ctx_kitchen_open.runner.push(_make_result(returncode=1))  # clone guard snapshot
-        tool_ctx_kitchen_open.runner.push(
-            _make_result(
-                0,
-                '{"type": "result", "subtype": "success", "is_error": false, '
-                '"result": "done", "session_id": "s1"}',
-                "",
-            )
-        )
-        await run_skill("/investigate error", "/tmp")
-        cmd = tool_ctx_kitchen_open.runner.call_args_list[-1][0]
-        prompt_idx = cmd.index("--print") + 1 if "--print" in cmd else cmd.index("-p") + 1
-        assert cmd[prompt_idx].startswith("Use the /investigate skill error")
-        actual_cwd = tool_ctx_kitchen_open.runner.call_args_list[-1][1]
-        assert actual_cwd == Path("/tmp").resolve(), (
-            f"Subprocess cwd mismatch: {actual_cwd} != {Path('/tmp').resolve()}"
-        )
-
-    @pytest.mark.anyio
-    async def test_run_skill_rejects_prose_without_slash(self, tool_ctx):
-        """FRICT-6-1: prose command without slash returns gate_error before reaching executor."""
-        result = json.loads(await run_skill("Fix the authentication bug in main.py", "/tmp"))
-        assert result["success"] is False
-        assert result["is_error"] is True
-        assert result["subtype"] == "gate_error"
-        # executor must NOT have been called
-        assert tool_ctx.runner.call_args_list == []
-
-    @pytest.mark.anyio
-    async def test_run_skill_rejects_empty_skill_command(self, tool_ctx):
-        """FRICT-6-1: empty string returns gate_error without hitting executor."""
-        result = json.loads(await run_skill("", "/tmp"))
-        assert result["success"] is False
-        assert result["is_error"] is True
-        assert result["subtype"] == "gate_error"
-        assert tool_ctx.runner.call_args_list == []
-
-    @pytest.mark.anyio
-    async def test_run_skill_rejects_whitespace_only(self, tool_ctx):
-        """FRICT-6-1: whitespace-only command returns gate_error (strip before check)."""
-        result = json.loads(await run_skill("   ", "/tmp"))
-        assert result["success"] is False
-        assert result["is_error"] is True
-        assert result["subtype"] == "gate_error"
-        assert tool_ctx.runner.call_args_list == []
-
-    @pytest.mark.anyio
-    async def test_run_skill_format_error_includes_slash_examples(self, tool_ctx_kitchen_open):
-        """FRICT-6-1: error message for invalid format includes concrete slash-command examples."""
-        result = json.loads(await run_skill("investigate this bug", "/tmp"))
-        assert result["success"] is False
-        assert "/autoskillit:" in result["result"]
-        assert "/" in result["result"]
-
-    @pytest.mark.anyio
-    async def test_run_skill_includes_completion_directive(self, tool_ctx_kitchen_open):
-        from tests.conftest import _make_result
-
-        tool_ctx_kitchen_open.runner.push(_make_result(returncode=1))  # clone guard snapshot
-        tool_ctx_kitchen_open.runner.push(
-            _make_result(
-                0,
-                '{"type": "result", "subtype": "success", "is_error": false, '
-                '"result": "done", "session_id": "s1"}',
-                "",
-            )
-        )
-        await run_skill("/investigate error", "/tmp")
-        cmd = tool_ctx_kitchen_open.runner.call_args_list[-1][0]
-        prompt_idx = cmd.index("--print") + 1 if "--print" in cmd else cmd.index("-p") + 1
-        assert "%%ORDER_UP::" in cmd[prompt_idx]
-        from pathlib import Path
-
-        actual_cwd = tool_ctx_kitchen_open.runner.call_args_list[-1][1]
-        assert actual_cwd == Path("/tmp").resolve(), (
-            f"Subprocess cwd mismatch: {actual_cwd} != {Path('/tmp').resolve()}"
-        )
-
-
-class TestValidateSkillCommand:
-    """Unit tests for _validate_skill_command helper."""
-
-    def test_returns_none_for_slash_command(self, tool_ctx):
-        from autoskillit.server._guards import _validate_skill_command
-
-        assert _validate_skill_command("/autoskillit:investigate") is None
-
-    def test_returns_none_for_bare_slash_command(self, tool_ctx):
-        from autoskillit.server._guards import _validate_skill_command
-
-        assert _validate_skill_command("/audit-arch") is None
-
-    def test_returns_error_json_for_prose(self, tool_ctx):
-        from autoskillit.server._guards import _validate_skill_command
-
-        result = _validate_skill_command("Fix the bug")
-        assert result is not None
-        parsed = json.loads(result)
-        assert parsed["success"] is False
-        assert parsed["subtype"] == "gate_error"
-
-    def test_returns_error_json_for_empty_string(self, tool_ctx):
-        from autoskillit.server._guards import _validate_skill_command
-
-        result = _validate_skill_command("")
-        assert result is not None
-
-    def test_strips_whitespace_before_check(self, tool_ctx):
-        from autoskillit.server._guards import _validate_skill_command
-
-        # Leading whitespace before slash → valid
-        assert _validate_skill_command("  /autoskillit:investigate") is None
-        # Leading whitespace before prose → invalid
-        result = _validate_skill_command("  investigate bug")
-        assert result is not None
-
-
-class TestDryWalkthroughGateWithPrefix:
-    """Dry-walkthrough gate still receives raw command before prefix is applied."""
-
-    @pytest.mark.anyio
-    async def test_gate_still_fires_for_implement_skill(self, tool_ctx_kitchen_open, tmp_path):
-        plan = tmp_path / "plan.md"
-        plan.write_text("# No marker plan")
-        result = json.loads(await run_skill(f"/implement-worktree {plan}", str(tmp_path)))
-        assert result["success"] is False
-        assert result["is_error"] is True
-        assert "dry-walked" in result["result"].lower()
-
-
-class TestInputContractValidation:
-    """_check_input_contracts validates file_path and directory_path inputs."""
-
-    def test_file_path_input_rejects_nonexistent_path(self, tmp_path):
-        from autoskillit.server._guards import _check_input_contracts
-
-        resolver = _make_input_contract_resolver()
-        result = _check_input_contracts(
-            "/resolve-failures /nonexistent/worktree /nonexistent/plan.md main",
-            str(tmp_path),
-            resolver,
-        )
-        assert result is not None
-        parsed = json.loads(result)
-        assert parsed["success"] is False
-        assert parsed["subtype"] == "gate_error"
-
-    def test_file_path_input_rejects_directory_as_file(self, tmp_path):
-        from autoskillit.server._guards import _check_input_contracts
-
-        worktree = tmp_path / "worktree"
-        worktree.mkdir()
-        plan_as_dir = tmp_path / "plan.md"
-        plan_as_dir.mkdir()
-        resolver = _make_input_contract_resolver()
-        result = _check_input_contracts(
-            f"/resolve-failures {worktree} {plan_as_dir} main",
-            str(tmp_path),
-            resolver,
-        )
-        assert result is not None
-        parsed = json.loads(result)
-        assert parsed["success"] is False
-
-    def test_file_path_input_accepts_file_without_extension(self, tmp_path):
-        from autoskillit.server._guards import _check_input_contracts
-
-        worktree = tmp_path / "worktree"
-        worktree.mkdir()
-        plan_no_ext = tmp_path / "plan_no_ext"
-        plan_no_ext.write_text("content")
-        resolver = _make_input_contract_resolver()
-        result = _check_input_contracts(
-            f"/resolve-failures {worktree} {plan_no_ext} main",
-            str(tmp_path),
-            resolver,
-        )
-        # plan_no_ext exists as a file so is_file() passes — this tests the file exists check
-        assert result is None
-
-    def test_directory_path_input_rejects_file_as_directory(self, tmp_path):
-        from autoskillit.server._guards import _check_input_contracts
-
-        file_as_worktree = tmp_path / "worktree"
-        file_as_worktree.write_text("I am a file")
-        resolver = _make_input_contract_resolver()
-        result = _check_input_contracts(
-            f"/resolve-failures {file_as_worktree} /some/plan.md main",
-            str(tmp_path),
-            resolver,
-        )
-        assert result is not None
-        parsed = json.loads(result)
-        assert parsed["success"] is False
-
-    def test_directory_path_input_accepts_valid_directory(self, tmp_path):
-        from autoskillit.server._guards import _check_input_contracts
-
-        worktree = tmp_path / "worktree"
-        worktree.mkdir()
-        plan = tmp_path / "plan.md"
-        plan.write_text("content")
-        resolver = _make_input_contract_resolver()
-        result = _check_input_contracts(
-            f"/resolve-failures {worktree} {plan} main",
-            str(tmp_path),
-            resolver,
-        )
-        assert result is None
-
-    def test_file_path_input_accepts_valid_file(self, tmp_path):
-        from autoskillit.server._guards import _check_input_contracts
-
-        worktree = tmp_path / "worktree"
-        worktree.mkdir()
-        plan = tmp_path / "my-plan.md"
-        plan.write_text("content")
-        resolver = _make_input_contract_resolver()
-        result = _check_input_contracts(
-            f"/resolve-failures {worktree} {plan} main",
-            str(tmp_path),
-            resolver,
-        )
-        assert result is None
-
-    def test_skill_without_contracts_passes(self, tmp_path):
-        from autoskillit.server._guards import _check_input_contracts
-
-        resolver = _make_input_contract_resolver()
-        result = _check_input_contracts(
-            "/nonexistent-skill-not-in-contracts /some/path.md",
-            str(tmp_path),
-            resolver,
-        )
-        assert result is None
-
-    def test_mangled_plan_path_with_timestamp_suffix(self, tmp_path):
-        from autoskillit.server._guards import _check_input_contracts
-
-        real_plan = tmp_path / "rectify_foo_2026-05-28_194500.md"
-        real_plan.write_text("Dry-walkthrough verified = TRUE")
-        worktree = tmp_path / "worktree"
-        worktree.mkdir()
-        mangled = tmp_path / "rectify_foo_2026-05-28_194500-20260528-201752"
-        resolver = _make_input_contract_resolver()
-        result = _check_input_contracts(
-            f"/resolve-failures {worktree} {mangled} main",
-            str(tmp_path),
-            resolver,
-        )
-        assert result is not None
-        parsed = json.loads(result)
-        assert parsed["success"] is False
-
-    def test_resolver_returns_none_skips_validation(self, tmp_path):
-        from autoskillit.server._guards import _check_input_contracts
-
-        result = _check_input_contracts(
-            "/resolve-failures /nonexistent/worktree /nonexistent/plan.md main",
-            str(tmp_path),
-            None,
-        )
-        assert result is None
-
-
-class TestInputContractResolver:
-    """InputContractResolver loads input specs from skill_contracts.yaml."""
-
-    def test_input_contract_resolver_returns_specs_for_resolve_failures(self):
-        from autoskillit.core import InputSpec
-
-        resolver = _make_input_contract_resolver()
-        specs = resolver("/resolve-failures /worktrees/foo /plans/bar.md main")
-        assert len(specs) == 3
-        assert specs[0] == InputSpec(
-            name="worktree_path", type="directory_path", required=True, position=0
-        )
-        assert specs[1] == InputSpec(name="plan_path", type="file_path", required=True, position=1)
-        assert specs[2] == InputSpec(
-            name="diagnosis_path", type="file_path", required=False, position=2
-        )
-
-    def test_input_contract_resolver_returns_empty_for_unknown_skill(self):
-        resolver = _make_input_contract_resolver()
-        specs = resolver("/unknown-skill-not-in-contracts /some/path")
-        assert list(specs) == []
-
-
-def _make_input_contract_resolver():
-    """Create a concrete InputContractResolver using the bundled manifest."""
-    from autoskillit.recipe._contracts_manifest import resolve_input_specs
-
-    return resolve_input_specs
-
-
-def _collect_all_path_input_specs() -> list[tuple[str, str, str]]:
-    from autoskillit.core.io import load_yaml
-
-    yaml_path = Path(__file__).parents[2] / "src/autoskillit/recipe/skill_contracts.yaml"
-    raw = load_yaml(yaml_path)
-    result = []
-    for skill_name, contract in sorted(raw.get("skills", {}).items()):
-        for inp in contract.get("inputs", []):
-            if inp.get("type") in ("file_path", "directory_path", "file_path_list"):
-                result.append((skill_name, inp["name"], inp["type"]))
-    return result
-
-
-_ALL_PATH_INPUT_SPECS = _collect_all_path_input_specs()
-
-
-def _collect_file_path_list_specs() -> list[tuple[str, str, str]]:
-    return [(s, i, t) for (s, i, t) in _ALL_PATH_INPUT_SPECS if t == "file_path_list"]
-
-
-_FILE_PATH_LIST_SPECS = _collect_file_path_list_specs()
-
-
-class TestInputContractRealContracts:
-    """Gate integration tests using real contract declarations."""
-
-    @pytest.mark.parametrize(
-        "skill_name,input_name,declared_type",
-        _ALL_PATH_INPUT_SPECS,
-        ids=[f"{s}-{i}" for s, i, _ in _ALL_PATH_INPUT_SPECS],
-    )
-    def test_gate_accepts_correct_path_type(
-        self, tmp_path, skill_name: str, input_name: str, declared_type: str
-    ):
-        """Gate must accept the correct filesystem entity for every declared path input."""
-        from autoskillit.core import InputSpec
-        from autoskillit.server._guards import _check_input_contracts
-
-        narrowed = cast("Literal['file_path', 'directory_path', 'file_path_list']", declared_type)
-        if narrowed == "file_path":
-            target = tmp_path / "test_input.md"
-            target.write_text("test")
-            cmd_value = str(target)
-        elif narrowed == "directory_path":
-            target = tmp_path / "test_input_dir"
-            target.mkdir()
-            cmd_value = str(target)
-        else:
-            member_a = tmp_path / "test_input_a.md"
-            member_a.write_text("a")
-            member_b = tmp_path / "test_input_b.md"
-            member_b.write_text("b")
-            cmd_value = f"{member_a},{member_b}"
-        spec = InputSpec(name=input_name, type=narrowed, required=True, position=0)
-        result = _check_input_contracts(
-            f"/autoskillit:{skill_name} {cmd_value}",
-            str(tmp_path),
-            resolver=lambda skill_command, _s=spec: (_s,),
-        )
-        assert result is None
-
-    @pytest.mark.parametrize(
-        "skill_name,input_name,declared_type",
-        _ALL_PATH_INPUT_SPECS,
-        ids=[f"{s}-{i}" for s, i, _ in _ALL_PATH_INPUT_SPECS],
-    )
-    def test_gate_rejects_wrong_path_type(
-        self, tmp_path, skill_name: str, input_name: str, declared_type: str
-    ):
-        """Gate must reject the WRONG filesystem entity for every declared path input."""
-        from autoskillit.core import InputSpec
-        from autoskillit.server._guards import _check_input_contracts
-
-        narrowed = cast("Literal['file_path', 'directory_path', 'file_path_list']", declared_type)
-        if narrowed == "file_path":
-            wrong_target = tmp_path / "wrong_dir"
-            wrong_target.mkdir()
-            cmd_value = str(wrong_target)
-        elif narrowed == "directory_path":
-            wrong_target = tmp_path / "wrong_file.md"
-            wrong_target.write_text("test")
-            cmd_value = str(wrong_target)
-        else:
-            existing = tmp_path / "existing_member.md"
-            existing.write_text("ok")
-            missing = tmp_path / "missing_member.md"
-            cmd_value = f"{existing},{missing}"
-        spec = InputSpec(name=input_name, type=narrowed, required=True, position=0)
-        result = _check_input_contracts(
-            f"/autoskillit:{skill_name} {cmd_value}",
-            str(tmp_path),
-            resolver=lambda skill_command, _s=spec: (_s,),
-        )
-        assert result is not None
-        parsed = json.loads(result)
-        assert parsed["success"] is False
-
-
-class TestFilePathListRealResolver:
-    """Real-resolver coverage for declared file_path_list specs."""
-
-    @pytest.mark.parametrize(
-        "skill_name,input_name,declared_type",
-        _FILE_PATH_LIST_SPECS,
-        ids=[f"{s}-{i}" for s, i, _ in _FILE_PATH_LIST_SPECS],
-    )
-    def test_gate_real_resolver_for_file_path_list_specs(
-        self, tmp_path, skill_name: str, input_name: str, declared_type: str
-    ) -> None:
-        """For every file_path_list spec, the real resolver must validate comma-joined members."""
-        from autoskillit.recipe._contracts_manifest import resolve_input_specs
-        from autoskillit.server._guards import _check_input_contracts
-
-        assert declared_type == "file_path_list", "parametrize filter precondition"
-        specs = resolve_input_specs(f"/autoskillit:{skill_name}")
-        target_spec = next((s for s in specs if s.name == input_name), None)
-        assert target_spec is not None, (
-            f"resolver did not surface {skill_name}.{input_name} (specs={specs!r})"
-        )
-
-        cmd_args: list[str] = []
-        for spec in specs:
-            if spec.position < target_spec.position:
-                if spec.type == "file_path":
-                    predecessor = tmp_path / f"pre_{spec.name}.md"
-                    predecessor.write_text("pre")
-                    cmd_args.append(str(predecessor))
-                elif spec.type == "directory_path":
-                    pred_dir = tmp_path / f"pre_{spec.name}_dir"
-                    pred_dir.mkdir()
-                    cmd_args.append(str(pred_dir))
-                else:
-                    a = tmp_path / f"pre_{spec.name}_a.md"
-                    a.write_text("a")
-                    b = tmp_path / f"pre_{spec.name}_b.md"
-                    b.write_text("b")
-                    cmd_args.append(f"{a},{b}")
-        member_a = tmp_path / f"target_{input_name}_a.md"
-        member_a.write_text("a")
-        member_b = tmp_path / f"target_{input_name}_b.md"
-        member_b.write_text("b")
-        cmd_args.append(f"{member_a},{member_b}")
-
-        cmd = f"/autoskillit:{skill_name} {' '.join(cmd_args)}"
-        result = _check_input_contracts(cmd, str(tmp_path), resolve_input_specs)
-        assert result is None, (
-            f"Expected gate acceptance for {skill_name}.{input_name}, got: {result!r}"
-        )
-
-
-class TestPreparePrCrossBinding:
-    """Bundled prepare-pr cross-binding: comma-joined plans must bind plan_paths at position 0."""
-
-    def test_prepare_pr_actual_recipe_shape_accepts_comma_plan_paths(self, tmp_path) -> None:
-        """Bundled recipe shape with comma-joined plans must clear the gate."""
-        from autoskillit.recipe._contracts_manifest import resolve_input_specs
-        from autoskillit.server._guards import _check_input_contracts
-
-        plan_a = tmp_path / "plan_a.md"
-        plan_a.write_text("plan a")
-        plan_b = tmp_path / "plan_b.md"
-        plan_b.write_text("plan b")
-        joined = f"{plan_a},{plan_b}"
-        cmd = f"/autoskillit:prepare-pr {joined} my-run main 123 true"
-        result = _check_input_contracts(cmd, str(tmp_path), resolve_input_specs)
-        assert result is None, f"Expected acceptance, got: {result!r}"
-
-    def test_prepare_pr_plan_list_and_conflict_report_bind_separately(self, tmp_path) -> None:
-        """Comma-joined plans plus optional conflict-report must bind as two path specs."""
-        from autoskillit.recipe._contracts_manifest import resolve_input_specs
-        from autoskillit.server._guards import _check_input_contracts
-
-        plan_a = tmp_path / "plan_a.md"
-        plan_a.write_text("plan a")
-        plan_b = tmp_path / "plan_b.md"
-        plan_b.write_text("plan b")
-        conflict = tmp_path / "conflict.md"
-        conflict.write_text("conflict")
-        joined = f"{plan_a},{plan_b}"
-        cmd = f"/autoskillit:prepare-pr {joined} my-run main 123 true {conflict}"
-        result = _check_input_contracts(cmd, str(tmp_path), resolve_input_specs)
-        assert result is None, f"Expected acceptance, got: {result!r}"
-
-    def test_prepare_pr_missing_conflict_report_is_rejected(self, tmp_path) -> None:
-        """When conflict-report is missing, the scalar spec at path position 1 fails."""
-        from autoskillit.recipe._contracts_manifest import resolve_input_specs
-        from autoskillit.server._guards import _check_input_contracts
-
-        plan_a = tmp_path / "plan_a.md"
-        plan_a.write_text("plan a")
-        plan_b = tmp_path / "plan_b.md"
-        plan_b.write_text("plan b")
-        joined = f"{plan_a},{plan_b}"
-        cmd = f"/autoskillit:prepare-pr {joined} my-run main 123 true /tmp/does-not-exist.md"
-        result = _check_input_contracts(cmd, str(tmp_path), resolve_input_specs)
-        assert result is not None
-        parsed = json.loads(result)
-        assert parsed["success"] is False
-        assert "conflict_report_path" in parsed["result"]
-
-
-class TestOpenIntegrationPrExactShape:
-    """Bundled open-integration-pr shape: conflict_report_paths must bind path position 1."""
-
-    def _make_recipe_shape_files(self, tmp_path: Path) -> tuple[Path, Path, Path, Path]:
-        pr_order = tmp_path / "pr_order.json"
-        pr_order.write_text("{}")
-        conflict_a = tmp_path / "conflict_a.md"
-        conflict_a.write_text("a")
-        conflict_b = tmp_path / "conflict_b.md"
-        conflict_b.write_text("b")
-        domain_partitions = tmp_path / "domain_partitions.json"
-        domain_partitions.write_text("{}")
-        return pr_order, conflict_a, conflict_b, domain_partitions
-
-    def test_open_integration_pr_recipe_shape_accepts_conflict_list(self, tmp_path: Path) -> None:
-        """Comma-joined conflict reports + named domain partitions must validate."""
-        from autoskillit.recipe._contracts_manifest import resolve_input_specs
-        from autoskillit.server._guards import _check_input_contracts
-
-        pr_order, conflict_a, conflict_b, domain_partitions = self._make_recipe_shape_files(
-            tmp_path
-        )
-        cmd = (
-            f"/autoskillit:open-integration-pr batch-branch main {pr_order} GO "
-            f'"{conflict_a},{conflict_b}" domain_partitions_path={domain_partitions}'
-        )
-        result = _check_input_contracts(cmd, str(tmp_path), resolve_input_specs)
-        assert result is None, f"Expected acceptance, got: {result!r}"
-
-    def test_open_integration_pr_missing_conflict_member_rejected(self, tmp_path: Path) -> None:
-        """Missing conflict-report member must be rejected as conflict_report_paths."""
-        from autoskillit.recipe._contracts_manifest import resolve_input_specs
-        from autoskillit.server._guards import _check_input_contracts
-
-        pr_order, conflict_a, _conflict_b, domain_partitions = self._make_recipe_shape_files(
-            tmp_path
-        )
-        missing = tmp_path / "missing_conflict.md"
-        cmd = (
-            f"/autoskillit:open-integration-pr batch-branch main {pr_order} GO "
-            f'"{conflict_a},{missing}" domain_partitions_path={domain_partitions}'
-        )
-        result = _check_input_contracts(cmd, str(tmp_path), resolve_input_specs)
-        assert result is not None
-        parsed = json.loads(result)
-        assert parsed["success"] is False
-        assert "conflict_report_paths" in parsed["result"]
-
-
-class TestInputContractIntegration:
-    """_check_input_contracts fires through run_skill when resolver is wired."""
-
-    @pytest.mark.anyio
-    async def test_run_skill_rejects_nonexistent_path_via_input_contract(
-        self, tool_ctx_kitchen_open
-    ):
-        tool_ctx_kitchen_open.input_contract_resolver = _make_input_contract_resolver()
-        result = json.loads(
-            await run_skill(
-                "/resolve-failures /nonexistent/worktree /nonexistent/plan.md main",
-                "/tmp",
-            )
-        )
-        assert result["success"] is False
-        assert result["subtype"] == "gate_error"
-        assert "Missing required" in result["result"] or "does not exist" in result["result"]
-        assert tool_ctx_kitchen_open.runner.call_args_list == []
-
-
-class TestRunSkillCwdValidation:
-    """run_skill rejects non-empty relative cwd at the boundary."""
-
-    @pytest.mark.anyio
-    async def test_run_skill_rejects_relative_cwd(self, tool_ctx_kitchen_open):
-        """Non-empty relative cwd is rejected immediately with a clear diagnostic."""
-        result = json.loads(
-            await run_skill(
-                "/autoskillit:retry-worktree plan.md ../worktrees/impl-fix",
-                cwd="../worktrees/impl-fix-20260316",
-            )
-        )
-        assert result["success"] is False
-        assert "cwd must be an absolute path" in result["error"]
-        assert "../worktrees/impl-fix-20260316" in result["error"]
-        assert tool_ctx_kitchen_open.runner.call_args_list == []
-
-    @pytest.mark.anyio
-    async def test_run_skill_accepts_empty_cwd(self, tool_ctx_kitchen_open, monkeypatch):
-        """Empty cwd is accepted (some skills have no specific cwd requirement)."""
-        from tests.conftest import _make_result
-
-        _patch_uuid4(monkeypatch)
-        marker = _DETERMINISTIC_MARKER
-        success_json = (
-            '{"type": "result", "subtype": "success", "is_error": false,'
-            f' "result": "done {marker}", "session_id": "s1"}}'
-        )
-        tool_ctx_kitchen_open.runner.push(_make_result(returncode=0, stdout=success_json))
-        result = json.loads(await run_skill("/investigate foo", cwd=""))
-        assert "cwd must be an absolute path" not in result.get("error", "")
-        assert result.get("subtype") != "gate_error"
-
-    @pytest.mark.anyio
-    async def test_run_skill_accepts_absolute_cwd(self, tool_ctx_kitchen_open, monkeypatch):
-        """Absolute cwd passes the boundary check and proceeds normally."""
-        from tests.conftest import _make_result
-
-        _patch_uuid4(monkeypatch)
-        marker = _DETERMINISTIC_MARKER
-        success_json = (
-            '{"type": "result", "subtype": "success", "is_error": false,'
-            f' "result": "done {marker}", "session_id": "s1"}}'
-        )
-        tool_ctx_kitchen_open.runner.push(_make_result(returncode=1))  # clone guard snapshot
-        tool_ctx_kitchen_open.runner.push(_make_result(returncode=0, stdout=success_json))
-        result = json.loads(await run_skill("/investigate foo", cwd="/tmp"))
-        assert "cwd must be an absolute path" not in result.get("error", "")
-        assert result.get("subtype") != "gate_error"
